Move skills to skills package
